backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=43.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.12
websockets>=12.0
aiofiles>=24.0.0
mcrcon>=0.7.0
httpx>=0.27.0
pydantic>=2.0.0
psutil>=5.9.0
docker>=7.0.0
